// src/app/models/card.ts
 export class Card {
        constructor(public Id: string,
            public Name: string,
            public Text: string,
            public Flavor: string,
            public Artist: string,
            public Attack: number,
            public Health: number,
            public Cost: number,
            public Rarity: Rarity,
            public CardClass: CardClass,
            public Collectible: string,
            public Elite: boolean,
            public Faction: Faction,
            public Mechanics: string[],            
            public Set: CardSet,
            public Type: CardType) {
        }
    }
// ...
   enum CardClass {
       INVALID = 0,
       DEATHKNIGHT = 1,
       DRUID = 2,
       HUNTER = 3,
       MAGE = 4,
       PALADIN = 5,
       PRIEST = 6,
       ROGUE = 7,
       SHAMAN = 8,
       WARLOCK = 9,
       WARRIOR = 10,
       DREAM = 11,
       NEUTRAL = 12,
   }

  export enum CardSet {
       INVALID = 0,
       TEST_TEMPORARY = 1,
       CORE = 2,
       EXPERT1 = 3,
       REWARD = 4,
       MISSIONS = 5,
       DEMO = 6,
       NONE = 7,
       CHEAT = 8,
       BLANK = 9,
       DEBUG_SP = 10,
       PROMO = 11,
       NAXX = 12,
       FP1 = 12,
       PE1 = 13,
       GVG = 13,
       BRM = 14,
       FP2 = 14,
       PE2 = 15,
       TEMP1 = 15,
       TGT = 15,
       CREDITS = 16,
       HERO_SKINS = 17,
       TB = 18,
       SLUSH = 19,
       LOE = 20,
       OG = 21,
       OG_RESERVE = 22,
       KARA = 23,
       KARA_RESERVE = 24,
       GANGS = 25,
       GANGS_RESERVE = 26,
       UNGORO = 27,
   }

export enum CardType {
       INVALID = 0,
       GAME = 1,
       PLAYER = 2,
       HERO = 3,
       MINION = 4,
       ABILITY = 5,
       SPELL = 5,
       ENCHANTMENT = 6,
       WEAPON = 7,
       ITEM = 8,
       TOKEN = 9,
       HERO_POWER = 10,
   }

   enum ChoiceType {
       INVALID = 0,
       MULLIGAN = 1,
       GENERAL = 2,
   }

   enum DeckType {
       NORMAL_DECK = 1,
       AI_DECK = 2,
       DRAFT_DECK = 4,
       PRECON_DECK = 5,
       TAVERN_BRAWL_DECK = 6,
       HIDDEN_DECK = 1000,
   }

   enum EnchantmentVisual {
       INVALID = 0,
       POSITIVE = 1,
       NEGATIVE = 2,
       NEUTRAL = 3,
   }

   enum Faction {
       INVALID = 0,
       HORDE = 1,
       ALLIANCE = 2,
       NEUTRAL = 3,
   }
// ...
   enum Rarity {
       INVALID = 0,
       COMMON = 1,
       FREE = 2,
       RARE = 3,
       EPIC = 4,
       LEGENDARY = 5,
       UNKNOWN_6 = 6,
   }
// ...
   enum Type {
       LOCSTRING = -2,
       UNKNOWN = 0,
       BOOL = 1,
       NUMBER = 2,
       COUNTER = 3,
       ENTITY = 4,
       PLAYER = 5,
       TEAM = 6,
       ENTITY_DEFINITION = 7,
       STRING = 8,
   }